chore: move service schemas to separate module
